Refactor. refactor Listview Table into several components
